[GDI32/WIN32K] - Rewrite GetCurrentObject and GetObject, fixes a lot of tests
svn path=/trunk/; revision=56400

+++ reactos/win32ss/gdi/gdi32/misc/stubs.c
@@ -1131,7 +1131,7 @@ GetETM(HDC hdc,
     BOOL Ret = NtGdiGetETM(hdc, petm);
 
     if (Ret && petm)
-        petm->emKernPairs = GetKerningPairsA(hdc, 0, 0);
+        petm->emKernPairs = (WORD)GetKerningPairsA(hdc, 0, 0);
 
     return Ret;
 }
